version 0.2.4 * Began code transfer/reuse from src-expresskeys to src-expresskeysconf as a scaffold for some kind of graphical utility. Long way to go...
* Explained the X "xev" program and keycodes a little better in
the USAGE file.

* Put a warning regarding the X program "xmodmap" in the BUGS and
in the USAGE file.

* Cleaned up gcc warnings on a x86_64 system.

_Data Type Sizes (for my own reference)_

Type		x86	x86_64

char		8	8
short		16	16
int		32	32
long		32	64
long long	64	64
float		32	32
double		64	64
long double	96	128
void*		32	64

In short, a programmer should NOT assume that:
sizeof(int) == sizeof(long) == sizeof(void*)

+++ src-expresskeys/config_write.c
@@ -79,7 +79,9 @@ int write_file_config(int *ip, FILE *fp)
 {
 
 	struct program *p;
-	p = (void *)*ip;
+	
+	/* Convert to long for x86_64 systems */
+	p = (void *)(long)*ip;
 
 	fprintf(fp, "%s				# <--- Begin New Program Record\n\n", "%%");
 	fprintf(fp, "00 Program Name: \"%s\"	# Name must be within double quotes \"\"\n", p->class_name);
